+++ cosmos/airflow/graph.py
@@ -18,46 +18,6 @@ from cosmos.render import calculate_operator_class
 logger = logging.getLogger(__name__)
 
 
-def filter_nodes(project_dir, nodes, select, exclude):
-    chosen_nodes = nodes.copy()
-
-    def build_paths(selected_paths):
-        return [(project_dir / path.strip("/")) for path in selected_paths]
-
-    SUPPORTED_CONFIG = ["materialized", "schema", "tags"]
-
-    def normalize_config(config):
-        values = []
-        for key, value in config.items():
-            if key in SUPPORTED_CONFIG:
-                if isinstance(value, list):
-                    normalized = [f"{key}:{item}" for item in value]
-                    values.extend(normalized)
-                else:
-                    values.append(f"{key}:{value}")
-        return values
-
-    for node_id, node in nodes.items():
-        # path-based filters
-        paths_to_include = build_paths(select.get("paths", []))
-        paths_to_exclude = build_paths(exclude.get("paths", []))
-        node_parent_paths = node.file_path.parents()
-        if not set(paths_to_include).intersection(node_parent_paths) or set(paths_to_exclude).intersection(
-            node_parent_paths
-        ):
-            chosen_nodes.pop(node_id)
-
-        # config-based filters
-        configs_to_include = normalize_config(select.get("configs", []))
-        configs_to_exclude = normalize_config(exclude.get("configs", []))
-        # TODO: add tags
-        node_config = normalize_config(node.config)
-        if not set(configs_to_include).intersection(node_config) or set(configs_to_exclude).intersection(node_config):
-            chosen_nodes.pop(node_id)
-
-        return chosen_nodes
-
-
 def calculate_leaves(nodes):
     """
     Tasks which are not parents (dependencies) to other tasks.
@@ -104,7 +64,7 @@ def create_test_task_metadata(test_task_name, execution_mode, task_args, on_warn
     )
 
 
-def add_airflow_entities(
+def build_airflow_graph(
     nodes: list[DbtNode],
     dag: DAG,  # Airflow-specific - parent DAG where to associate tasks and (optional) task groups
     execution_mode: str,  # Cosmos-specific - decide what which class to use
